+++ lute/luau/CMakeLists.txt
@@ -5,11 +5,13 @@ target_sources(Lute.Luau PRIVATE
     include/lute/luau.h
     include/lute/moduleresolver.h
     include/lute/resolverequire.h
+    include/lute/type.h
 
     src/configresolver.cpp
     src/luau.cpp
     src/moduleresolver.cpp
     src/resolverequire.cpp
+    src/type.cpp
 )
 
 target_compile_features(Lute.Luau PUBLIC cxx_std_17)
